requirements.txt
pandas
matplotlib
requests
python-dotenv
pyTelegramBotAPI==4.15.4
requests==2.31.0
ccxt==4.2.59
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
google-genai==0.3.0
python-dotenv==1.0.1
schedule==1.2.1
Flask==3.0.2
claude_brain.py